Inline gateway lib into gateway bin
