--- rust/connlib/libs/gateway/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "firezone-gateway-connlib"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-libs-common = { path = "../common" }
-async-trait = { version = "0.1", default-features = false }
-firezone-tunnel = { path = "../tunnel" }
-tokio = { version = "1.32", default-features = false, features = ["sync"] }
-tracing = { workspace = true }
-serde = { version = "1.0", default-features = false, features = ["std", "derive"] }
-boringtun = { workspace = true }
-chrono = { workspace = true }
-backoff = { workspace = true }
-webrtc = "0.8"
-url = { version = "2.4.1", default-features = false }
-rand = { version = "0.8", default-features = false, features = ["std"] }
-tokio-tungstenite = { version = "0.19", default-features = false, features = ["connect", "handshake", "rustls-tls-webpki-roots"] }
-
-[dev-dependencies]
-serde_json = { version = "1.0", default-features = false, features = ["std"] }
